enable apple-mobile-web-app-capable
so it opens without the safari chrome.
is there a security issue in this? I mean you won't know if you're
getting phished in safari anyway, and if this encourages bookmarking
then that's safer. Besides, burner wallet is burner wallet.

status-bar-style is black instead of white
chose this over black-translucent because it made the scroll
easier on the eye. even with black-translucent, there was still a black
bar at the top, it wasn't really transulcent until scrolling and at that
point what's the point, it just looked weird. Would be cool if we could
make the status bar the off-blank theme but :shrug: