Restructure tags intertwined with straight text in JSX.
This is precursor to fix issues outputted by the debug
functionality of Prettier to try and reduce unexpected/
unwanted changes.

The main fixes here ensure a line with JSX <....> and text
preceeding it doesn't add a {' '} assuming there needs to
be space there. It technically would work (since it's trying
to cut down line length). However, the better option is to
be more explicit in what is needed. Wrap in string literal
with interpolation or add <span> tags around.

- Update all reported outputs from format:debug.